Spikes (Hazards) for player
Create a hazard object of some kind that destroys the cube player object and restarts the level if the player cube object is destroyed. E.g., a spike hazard that destroys the player object if touched and restarts the current level in some fashion (find some free assets on the unity asset store to add some flavor!).

Created animations for spikes (3d cubes). If they touch the player the game will end